agent-starter-python/requirements.txt
# Core runtime
python-dotenv==1.2.1
requests==2.32.5

# LiveKit agent stack
livekit==1.0.19
livekit-agents==1.3.3
livekit-api==1.0.7
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-silero==1.3.3
livekit-plugins-turn-detector==1.3.3

# LangChain + RAG stack
langchain==1.0.8
langchain-core==1.0.7
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-chroma==1.0.0
langchain-huggingface==1.0.1
chromadb==1.3.5
pypdf==6.3.0
sentence-transformers==5.1.2
scikit-learn==1.7.2
torch==2.9.1

# LLM / OpenAI client
openai==2.8.1
tiktoken==0.12.0

# Utilities
numpy==2.2.6
pydantic==2.12.4
pydantic-settings==2.12.0
rich==14.2.0
uvloop==0.22.1
sounddevice==0.5.3
